+++ src/Manager/Environment/OpenIde.jsx
