+++ src/Manager/Environment/ShowFirstSteps.jsx
@@ -34,62 +34,41 @@
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import './style.scss';
-
 import React from 'react';
 import { useDispatch } from 'react-redux';
-import { exec } from 'child_process';
-import path from 'path';
-import { shell } from 'electron';
-import Dropdown from 'react-bootstrap/Dropdown';
-import DropdownButton from 'react-bootstrap/DropdownButton';
-import {
-    cloneNcs,
-    confirmInstall,
-    confirmRemove,
-} from '../environmentsActions';
 
+import { gotoPage } from '../../launcherActions';
+import { selectEnvironment } from '../managerReducer';
+import Button from './Button';
 import environmentPropType from './environmentPropType';
 
-const openBash = toolchainDir => {
-    exec(`"${path.resolve(toolchainDir, 'git-bash.exe')}"`);
-};
-
-const openCmd = toolchainDir => {
-    exec(`start cmd /k "${path.resolve(toolchainDir, 'git-cmd.cmd')}"`);
-};
-
-const openFolder = folder => {
-    shell.openItem(folder);
-};
-
-const EnvironmentMenu = ({ environment: { isInProcess, toolchainDir, version } }) => {
+const ShowFirstSteps = ({
+    environment: {
+        toolchainDir,
+        version,
+        isRemoving,
+        isInProcess,
+    },
+}) => {
     const dispatch = useDispatch();
+
     const isInstalled = !!toolchainDir;
+    if (!isInstalled && !(isInProcess && !isRemoving)) return null;
 
     return (
-        <DropdownButton
-            className="ml-2"
+        <Button
+            icon="x-mdi-dog-service"
+            onClick={() => {
+                dispatch(selectEnvironment(version));
+                dispatch(gotoPage(2));
+            }}
+            label="First steps to build"
+            title="Show how to build a sample project"
             variant="outline-primary"
-            title=""
-            alignRight
-            disabled={isInProcess || !isInstalled}
-        >
-            {/* eslint-disable max-len */}
-            <Dropdown.Item onClick={() => openBash(toolchainDir)}>Open bash</Dropdown.Item>
-            <Dropdown.Item onClick={() => openCmd(toolchainDir)}>Open command prompt</Dropdown.Item>
-            <Dropdown.Divider />
-            <Dropdown.Item onClick={() => openFolder(path.dirname(toolchainDir))}>Open SDK folder</Dropdown.Item>
-            <Dropdown.Item onClick={() => openFolder(toolchainDir)}>Open toolchain folder</Dropdown.Item>
-            <Dropdown.Divider />
-            <Dropdown.Item onClick={() => cloneNcs(dispatch, version, toolchainDir)}>Update SDK</Dropdown.Item>
-            <Dropdown.Item onClick={() => confirmInstall(dispatch, version)}>Update toolchain</Dropdown.Item>
-            <Dropdown.Divider />
-            <Dropdown.Item onClick={() => confirmRemove(dispatch, version)}>Remove</Dropdown.Item>
-            {/* eslint-enable max-len */}
-        </DropdownButton>
+        />
     );
 };
-EnvironmentMenu.propTypes = { environment: environmentPropType.isRequired };
 
-export default EnvironmentMenu;
+ShowFirstSteps.propTypes = { environment: environmentPropType.isRequired };
+
+export default ShowFirstSteps;
