+++ src/Manager/OtherPlatformInstructions.jsx
