+++ src/Manager/segger.js
