+++ src/persistentStore.js
@@ -34,7 +34,6 @@
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-
 import Store from 'electron-store';
 import path from 'path';
 import os from 'os';
